用户登陆：整合文件，把功能移到AdminUserController 把LoginController、LogoutController等文件删除
