--- src/test/java/com/vanke/tydirium/AdminApplicationTests.java
@@ -1,16 +1,0 @@
-package com.vanke.tydirium;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class AdminApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
